requirements.txt
amqp==1.4.9
configargparse==0.10
celery==3.1.18
chardet==2.3.0
colorlog==2.6.0
cssselect==0.9.1
Django>=1.11,<1.12
django-jsonfield==0.9.15
djangorestframework>=3.8,<4
elasticsearch>=5.4.0,<5.5.0
html2text==2016.1.8
jdcal==1.0.1
kombu==3.0.33
openpyxl==2.3.0
pyparsing>=2.2.0,<2.3.0
PyPDF2==1.25.1
python-dateutil==2.4.2
Unidecode==0.04.19
urllib3==1.13.1
python-Levenshtein==0.12.0
GitPython==1.0.1
psycopg2==2.6.1
requests==2.9.1
django-celery>=3.2,<3.3
amqplib==1.0.2
pygments==2.1
lxml==3.5.0
django-extensions==1.5.9
django-filter>0.11.0
markdown==2.6.2
pyinotify==0.9.6

pdfminer3k==1.3.0
splinter

tblib
tqdm
django-redis==4.6.0
ruamel.yaml==0.12.18

dateparser==0.7,<0.8

anyjson
billiard==3.3.0.23
six
pytz
KafNafParserPy
nlpipe==0.30
rpy2>=2.8,<2.9
iso8601

django-formtools
django-hash-field

actionform
amcatclient>=3.4
filemagic>=1.6.0,<1.7
